@@ -22,7 +22,9 @@
     }
   },
   "require": {
-    "php": "^5.6 || ^7.0",
+    "php": "^7.2",
+    "ext-intl": "^1.1",
+    "ext-gettext": "^7.2",
     "zendframework/zend-stdlib": "^3.0"
   },
   "require-dev": {
