@@ -6,13 +6,17 @@
 
 namespace Zetta\ZendBootstrap\Hydrator\Strategy;
 
+use DateTime;
+use DateTimeInterface;
+use IntlDateFormatter;
 use Jenssegers\Date\Date;
 use Zend\Hydrator\Strategy\StrategyInterface;
+use Zend\Hydrator\Strategy\Exception;
 
 class DateStrategy implements StrategyInterface
 {
     /**
-     * @var \IntlDateFormatter
+     * @var IntlDateFormatter
      */
     protected $formatter;
 
@@ -21,12 +25,12 @@ class DateStrategy implements StrategyInterface
      */
     public function __construct()
     {
-        $this->formatter = new \IntlDateFormatter(null, \IntlDateFormatter::SHORT, \IntlDateFormatter::NONE);
+        $this->formatter = new IntlDateFormatter(null, IntlDateFormatter::SHORT, IntlDateFormatter::NONE);
     }
 
     /**
      * Get the DateStrategy formatter
-     * @return \IntlDateFormatter
+     * @return IntlDateFormatter
      */
     public function getFormatter()
     {
@@ -35,7 +39,7 @@ class DateStrategy implements StrategyInterface
 
     /**
      * Set the DateStrategy formatter
-     * @param \IntlDateFormatter $formatter
+     * @param IntlDateFormatter $formatter
      * @return DateStrategy
      */
     public function setFormatter($formatter)
@@ -49,7 +53,7 @@ class DateStrategy implements StrategyInterface
      */
     public function extract($value)
     {
-        if ($value != null) {
+        if ($value instanceof DateTimeInterface) {
             return $value->format($this->getDateFormat());
         }
 
@@ -61,31 +65,37 @@ class DateStrategy implements StrategyInterface
      */
     public function hydrate($value)
     {
-        if ($value instanceof \DateTime) {
+        if ($value === '' || $value === null || $value instanceof DateTimeInterface) {
             return $value;
         }
 
-        if (is_string($value)) {
-            return Date::createFromFormat($this->getDateFormat() . ' H:i', $value . ' 00:00');
+        if (!is_string($value)) {
+            throw new Exception\InvalidArgumentException(sprintf(
+                'Unable to hydrate. Expected null, string, or DateTimeInterface; %s was given.',
+                is_object($value) ? get_class($value) : gettype($value)
+            ));
         }
 
-        return $value;
+        $hydrated = Date::createFromFormat($this->getDateFormat() . ' H:i', $value . ' 00:00');
+
+        return $hydrated ?: $value;
     }
 
     /**
      * @return null|string|string[]
+     * @throws \Exception
      */
     private function getDateFormat()
     {
 
-        $patterns = array(
+        $patterns = [
             '/11\D21\D(1999|99)/',
             '/21\D11\D(1999|99)/',
             '/(1999|99)\D11\D21/',
-        );
-        $replacements = array('m/d/Y', 'd/m/Y', 'Y/m/d');
+        ];
+        $replacements = ['m/d/Y', 'd/m/Y', 'Y/m/d'];
 
-        $date = new \DateTime();
+        $date = new DateTime();
         $date->setDate(1999, 11, 21);
         return preg_replace($patterns, $replacements, $this->formatter->format($date));
     }
