@@ -15,16 +15,16 @@ class Referer extends AbstractPlugin
     /**
      * Get referer URL based on a route
      *
-     * @param  string $route RouteInterface name
-     * @param  array|Traversable $params Parameters to use in url generation, if any
-     * @param  array|bool $options RouteInterface-specific options to use in url generation, if any.
+     * @param string $route RouteInterface name
+     * @param array|Traversable $params Parameters to use in url generation, if any
+     * @param array|bool $options RouteInterface-specific options to use in url generation, if any.
      *                                                If boolean, and no fourth argument, used as $reuseMatchedParams.
-     * @param  bool $reuseMatchedParams Whether to reuse matched parameters
+     * @param bool $reuseMatchedParams Whether to reuse matched parameters
      *
-     * @throws \Zend\Mvc\Exception\RuntimeException
-     * @throws \Zend\Mvc\Exception\InvalidArgumentException
-     * @throws \Zend\Mvc\Exception\DomainException
      * @return string
+     * @throws Exception\InvalidArgumentException
+     * @throws Exception\DomainException
+     * @throws Exception\RuntimeException
      */
     public function fromRoute($route = null, $params = [], $options = [], $reuseMatchedParams = false)
     {
