@@ -7,17 +7,23 @@
 namespace Zetta\ZendBootstrap\View\Helper;
 
 use Traversable;
+use Zend\Http\Request;
 use Zend\Stdlib\ArrayUtils;
-use Zend\Stdlib\RequestInterface;
 use Zend\View\Exception;
+use Zend\View\Helper\AbstractHelper;
 use Zend\View\Helper\Url as UrlHelper;
 
-class Url extends UrlHelper
+class Url extends AbstractHelper
 {
     /**
-     * @var RequestInterface
+     * @var Request
      */
     protected $request;
+
+    /**
+     * @var UrlHelper
+     */
+    protected $url;
 
     /**
      * @var bool
@@ -26,12 +32,14 @@ class Url extends UrlHelper
 
     /**
      * Referer constructor.
-     * @param RequestInterface $request
+     * @param Request $request
+     * @param UrlHelper $url
      * @param array $config
      */
-    public function __construct(RequestInterface $request, $config = [])
+    public function __construct(Request $request, UrlHelper $url, $config = [])
     {
         $this->request = $request;
+        $this->url = $url;
 
         if (isset($config['reuse_query']) && !empty($config['reuse_query'])) {
             $this->setReuseQuery($config['reuse_query']);
@@ -39,16 +47,16 @@ class Url extends UrlHelper
     }
 
     /**
-     * @param  string $name Name of the route
-     * @param  array $params Parameters for the link
-     * @param  array|Traversable $options Options for the route
-     * @param  bool $reuseMatchedParams Whether to reuse matched parameters
+     * @param string $name Name of the route
+     * @param array $params Parameters for the link
+     * @param array|Traversable $options Options for the route
+     * @param bool $reuseMatchedParams Whether to reuse matched parameters
      * @return string Url                         For the link href attribute
      */
     public function __invoke($name = null, $params = [], $options = [], $reuseMatchedParams = false)
     {
         if (null === $this->request) {
-            throw new Exception\RuntimeException('No RequestInterface instance provided');
+            throw new Exception\RuntimeException('No Request instance provided');
         }
 
         if (3 == func_num_args() && is_bool($options)) {
@@ -68,28 +76,11 @@ class Url extends UrlHelper
             }
         }
 
-        return parent::__invoke($name, $params, $options, $reuseMatchedParams);
+        return $this->url->__invoke($name, $params, $options, $reuseMatchedParams);
     }
 
     /**
-     * @return RequestInterface
-     */
-    public function getRequest()
-    {
-        return $this->request;
-    }
-
-    /**
-     * @param RequestInterface $request
-     * @return Url
-     */
-    public function setRequest($request)
-    {
-        $this->request = $request;
-        return $this;
-    }
-
-    /**
+     * Get the Url reuseQuery
      * @return bool
      */
     public function isReuseQuery()
@@ -98,6 +89,7 @@ class Url extends UrlHelper
     }
 
     /**
+     * Set the Url reuseQuery
      * @param bool $reuseQuery
      * @return Url
      */
